PARTIAL: Split launching and stop into warmup+assign and expire+teardown
